Refactor data preparation into commands
