+++ spinta/types/model.py
@@ -1,5 +1,3 @@
-from spinta.backends.mongo import Mongo
-from spinta.backends.postgresql import PostgreSQL
 from spinta.commands import load, check, error, authorize, prepare
 from spinta.components import Context, Manifest, Model, Property
 from spinta.nodes import load_node
@@ -79,24 +77,6 @@ def check(context: Context, model: Model, data: dict):
 @prepare.register()
 def prepare(context: Context, model: Model, data: dict) -> dict:
     data = prepare(context, model, model.backend, data)
-    return data
-
-
-@prepare.register()
-def prepare(context: Context, model: Model, backend: PostgreSQL, data: dict) -> dict:
-    for name, prop in model.properties.items():
-        if name in data:
-            data_value = data[name]
-            data[name] = prop.type.prepare_for_postgres(data_value)
-    return data
-
-
-@prepare.register()
-def prepare(context: Context, model: Model, backend: Mongo, data: dict) -> dict:
-    for name, prop in model.properties.items():
-        if name in data:
-            data_value = data[name]
-            data[name] = prop.type.prepare_for_mongo(data_value)
     return data
